refactor: background music provider, core provider를 core/providers 폴더로 이동
