+++ lib/core/providers/background_music_provider.dart
