+++ lib/features/auth/providers/account_provider.dart
@@ -1,10 +1,10 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:mongbi_app/core/providers/core_providers.dart';
 import 'package:mongbi_app/features/auth/data/data_sources/account_data_source.dart';
 import 'package:mongbi_app/features/auth/data/data_sources/remote_account_data_source.dart';
 import 'package:mongbi_app/features/auth/data/dtos/remote_account_repository.dart';
 import 'package:mongbi_app/features/auth/domain/repositories/account_repository.dart';
 import 'package:mongbi_app/features/setting/domain/use_cases/remove_account_use_case.dart';
-import 'package:mongbi_app/providers/core_providers.dart';
 
 final _accountDataSourceProvider = Provider<AccountDataSource>((ref) {
   final dio = ref.read(dioProvider);
