+++ lib/features/setting/presentation/models/bgm_state.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mongbi_app/providers/background_music_provider.dart';
+import 'package:mongbi_app/core/providers/background_music_provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class BgmState extends Notifier<bool> {
